Big pipeline refactor
Reworked the way that Materials and Pipelines work

After some deliberation I removed the IPipeline and VKPipeline classes.
They seem fairly unnecessary. Pipelines are essentially systems that
describe shader variables and store shader modules. That sounds like what
a Material should be to me. Still need to have a way to properly build the
descriptor sets from the shader variables but that should be done later.