Brought back the Pipeline system
Turns out the Pipeline can't just be in a Material

The Pipeline can be used as a core part of our rendering system that must
be seperate from the Material. While a Material could store a Pipeline the
fact is that the Pipeline is like a GL shader program with the addition of
the raster state and other previous globals. If that is stuck in the
Material that means we have to have a shader program per material which is
limiting. If we did that we couldn't have two materials use the same
Pipeline (and its shaders) while having the material apply different
variables and textures.