VKMaterial work
The VKMaterial class is now a bit more fleshed out

It's not perfect but it can be prepared and even handles potential
textures pretty well. A lot of it will have to be handled in the init from
file step. The buffers are created properly but we have no way to properly
fill them unless we know what's going to be in them.