Began work on render target system
To get the render pass system to work properly we need offscreen rendering

The render passes need to render to offscreen textures to get the
architecture working the way we want. If we DON'T do that we need to tie
the render pass to the swapchain, which is ugly. The swapchain is really
the same command buffer rendering to two different framebuffers that get
swapped by the swapchain. We need to get the render pass to render off to
a texture and then give the render target of the main render pass to the
swapchain's command buffers. It's a whole rigamarole and it sucks. To get
rendering to the screen I might put this work down and just tie into the
swapchain but I'd like to avoid that.